+++ src/Actions/AdminActions.php
@@ -184,7 +184,7 @@ class AdminActions extends DefaultActions
         // parameter checking
         if (!preg_match('/^\w+$/u', $id)) {
             $this->log()->warning('admin_modify_idtemplate: invalid template id ' . $id);
-            return $this->responder->mkError(400, "Invalid ID for template: " . $id);
+            return $this->responder->error(400, "Invalid ID for template: " . $id);
         }
 
         $template_data = $this->requester->post();
@@ -203,11 +203,11 @@ class AdminActions extends DefaultActions
         }
         if (!is_null($ntemplate)) {
             $msg = $this->getMessageString('admin_modified');
-            $answer = json_encode(['template' => $ntemplate->unbox()->getProperties(), 'msg' => $msg]);
-            return $this->responder->mkResponse($answer, 'application/json', 200);
+            $data = ['template' => $ntemplate->unbox()->getProperties(), 'msg' => $msg];
+            return $this->responder->json($data);
         } else {
-            $answer = $this->getMessageString('admin_modify_error');
-            return $this->responder->mkResponse($answer, 'text/plain', 500);
+            $message = $this->getMessageString('admin_modify_error');
+            return $this->responder->error(500, $message);
         }
         #$this->log()->debug('admin_modify_idtemplate 2: '.var_export($ntemplate, true));
     }
@@ -221,18 +221,18 @@ class AdminActions extends DefaultActions
         // parameter checking
         if (!preg_match('/^\w+$/u', $id)) {
             $this->log()->warning('admin_clear_idtemplate: invalid template id ' . $id);
-            return $this->responder->mkError(400, "Invalid ID for template: " . $id);
+            return $this->responder->error(400, "Invalid ID for template: " . $id);
         }
 
         $this->log()->debug('admin_clear_idtemplate: ' . var_export($id, true));
         $success = $this->bbs()->deleteIdTemplate($id);
         if ($success) {
             $msg = $this->getMessageString('admin_modified');
-            $answer = json_encode(['msg' => $msg]);
-            return $this->responder->mkResponse($answer, 'application/json', 200);
+            $data = ['msg' => $msg];
+            return $this->responder->json($data);
         } else {
-            $answer = $this->getMessageString('admin_modify_error');
-            return $this->responder->mkResponse($answer, 'text/plain', 404);
+            $message = $this->getMessageString('admin_modify_error');
+            return $this->responder->error(404, $message);
         }
     }
 
@@ -318,7 +318,7 @@ class AdminActions extends DefaultActions
         // parameter checking
         if (!is_numeric($id)) {
             $this->log()->warning('admin_get_user: invalid user id ' . $id);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $user = $this->bbs()->user($id);
@@ -364,11 +364,11 @@ class AdminActions extends DefaultActions
         }
         if (isset($user)) {
             $msg = $this->getMessageString('admin_modified');
-            $answer = json_encode(['user' => $user->unbox()->getProperties(), 'msg' => $msg]);
-            return $this->responder->mkResponse($answer, 'application/json', 200);
+            $data = ['user' => $user->unbox()->getProperties(), 'msg' => $msg];
+            return $this->responder->json($data);
         } else {
-            $answer = $this->getMessageString('admin_modify_error');
-            return $this->responder->mkResponse($answer, 'text/plain', 500);
+            $message = $this->getMessageString('admin_modify_error');
+            return $this->responder->error(500, $message);
         }
     }
 
@@ -381,18 +381,18 @@ class AdminActions extends DefaultActions
         // parameter checking
         if (!is_numeric($id)) {
             $this->log()->warning('admin_delete_user: invalid user id ' . $id);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $this->log()->debug('admin_delete_user: ' . var_export($id, true));
         $success = $this->bbs()->deleteUser($id);
         if ($success) {
             $msg = $this->getMessageString('admin_modified');
-            $answer = json_encode(['msg' => $msg]);
-            return $this->responder->mkResponse($answer, 'application/json', 200);
+            $data = ['msg' => $msg];
+            return $this->responder->json($data);
         } else {
-            $answer = $this->getMessageString('admin_modify_error');
-            return $this->responder->mkResponse($answer, 'text/plain', 500);
+            $message = $this->getMessageString('admin_modify_error');
+            return $this->responder->error(500, $message);
         }
     }
 
@@ -405,7 +405,7 @@ class AdminActions extends DefaultActions
         // parameter checking
         if (!is_numeric($id)) {
             $this->log()->warning('admin_modify_user: invalid user id ' . $id);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $user_data = $this->requester->post();
@@ -420,11 +420,11 @@ class AdminActions extends DefaultActions
         $this->log()->debug('admin_modify_user: ' . json_encode($user));
         if (isset($user)) {
             $msg = $this->getMessageString('admin_modified');
-            $answer = json_encode(['user' => $user->unbox()->getProperties(), 'msg' => $msg]);
-            return $this->responder->mkResponse($answer, 'application/json', 200);
+            $data = ['user' => $user->unbox()->getProperties(), 'msg' => $msg];
+            return $this->responder->json($data);
         } else {
-            $answer = $this->getMessageString('admin_modify_error');
-            return $this->responder->mkResponse($answer, 'text/plain', 500);
+            $message = $this->getMessageString('admin_modify_error');
+            return $this->responder->error(500, $message);
         }
     }