+++ src/Actions/OpdsActions.php
@@ -73,7 +73,7 @@ class OpdsActions extends DefaultActions
     {
         $gen = $this->mkOpdsGenerator();
         $cat = $gen->rootCatalog(null);
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_NAV);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_NAV);
     }
 
     /**
@@ -100,7 +100,7 @@ class OpdsActions extends DefaultActions
         $books = array_map([$this, 'checkThumbnailOpds'], $books1);
         $gen = $this->mkOpdsGenerator();
         $cat = $gen->newestCatalog(null, $books, false);
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_ACQ);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_ACQ);
     }
 
     /**
@@ -119,7 +119,7 @@ class OpdsActions extends DefaultActions
         // parameter checking
         if (!is_numeric($page)) {
             $this->log()->warning('opdsByTitle: invalid page id ' . $page);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $filter = $this->getFilter();
@@ -140,7 +140,7 @@ class OpdsActions extends DefaultActions
             $this->getNextSearchPage($tl),
             $this->getLastSearchPage($tl)
         );
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_ACQ);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_ACQ);
     }
 
     /**
@@ -152,7 +152,7 @@ class OpdsActions extends DefaultActions
         $initials = $this->calibre()->authorsInitials();
         $gen = $this->mkOpdsGenerator();
         $cat = $gen->authorsRootCatalog(null, $initials);
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_NAV);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_NAV);
     }
 
     /**
@@ -165,13 +165,13 @@ class OpdsActions extends DefaultActions
         // parameter checking
         if (!(ctype_upper($initial))) {
             $this->log()->warning('opdsByAuthorNamesForInitial: invalid initial ' . $initial);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $authors = $this->calibre()->authorsNamesForInitial($initial);
         $gen = $this->mkOpdsGenerator();
         $cat = $gen->authorsNamesForInitialCatalog(null, $authors, $initial);
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_NAV);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_NAV);
     }
 
     /**
@@ -188,7 +188,7 @@ class OpdsActions extends DefaultActions
         // parameter checking
         if (!is_numeric($id) || !is_numeric($page)) {
             $this->log()->warning('opdsByAuthor: invalid author id ' . $id . ' or page id ' . $page);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $filter = $this->getFilter();
@@ -208,7 +208,7 @@ class OpdsActions extends DefaultActions
             $this->getNextSearchPage($tl),
             $this->getLastSearchPage($tl)
         );
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_ACQ);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_ACQ);
     }
 
     /**
@@ -220,7 +220,7 @@ class OpdsActions extends DefaultActions
         $initials = $this->calibre()->tagsInitials();
         $gen = $this->mkOpdsGenerator();
         $cat = $gen->tagsRootCatalog(null, $initials);
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_NAV);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_NAV);
     }
 
     /**
@@ -233,13 +233,13 @@ class OpdsActions extends DefaultActions
         // parameter checking
         if (!(ctype_upper($initial))) {
             $this->log()->warning('opdsByTagNamesForInitial: invalid initial ' . $initial);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $tags = $this->calibre()->tagsNamesForInitial($initial);
         $gen = $this->mkOpdsGenerator();
         $cat = $gen->tagsNamesForInitialCatalog(null, $tags, $initial);
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_NAV);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_NAV);
     }
 
     /**
@@ -256,7 +256,7 @@ class OpdsActions extends DefaultActions
         // parameter checking
         if (!is_numeric($id) || !is_numeric($page)) {
             $this->log()->warning('opdsByTag: invalid tag id ' . $id . ' or page id ' . $page);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $filter = $this->getFilter();
@@ -274,7 +274,7 @@ class OpdsActions extends DefaultActions
             $this->getNextSearchPage($tl),
             $this->getLastSearchPage($tl)
         );
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_ACQ);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_ACQ);
     }
 
     /**
@@ -286,7 +286,7 @@ class OpdsActions extends DefaultActions
         $initials = $this->calibre()->seriesInitials();
         $gen = $this->mkOpdsGenerator();
         $cat = $gen->seriesRootCatalog(null, $initials);
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_NAV);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_NAV);
     }
 
     /**
@@ -299,13 +299,13 @@ class OpdsActions extends DefaultActions
         // parameter checking
         if (!($initial == 'all' || ctype_upper($initial))) {
             $this->log()->warning('opdsBySeriesNamesForInitial: invalid initial ' . $initial);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $tags = $this->calibre()->seriesNamesForInitial($initial);
         $gen = $this->mkOpdsGenerator();
         $cat = $gen->seriesNamesForInitialCatalog(null, $tags, $initial);
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_NAV);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_NAV);
     }
 
     /**
@@ -322,7 +322,7 @@ class OpdsActions extends DefaultActions
         // parameter checking
         if (!is_numeric($id) || !is_numeric($page)) {
             $this->log()->warning('opdsBySeries: invalid series id ' . $id . ' or page id ' . $page);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $filter = $this->getFilter();
@@ -340,7 +340,7 @@ class OpdsActions extends DefaultActions
             $this->getNextSearchPage($tl),
             $this->getLastSearchPage($tl)
         );
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_ACQ);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_ACQ);
     }
 
     /**
@@ -351,7 +351,7 @@ class OpdsActions extends DefaultActions
     {
         $gen = $this->mkOpdsGenerator();
         $cat = $gen->searchDescriptor(null, '/opds/searchlist/0/');
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPENSEARCH_MIME);
+        return $this->responder->opds($cat, OpdsGenerator::OPENSEARCH_MIME);
     }
 
     /**
@@ -368,14 +368,14 @@ class OpdsActions extends DefaultActions
         // parameter checking
         if (!is_numeric($page)) {
             $this->log()->warning('opdsBySearch: invalid page id ' . $page);
-            return $this->responder->mkError(400, "Bad parameter");
+            return $this->responder->error(400, "Bad parameter");
         }
 
         $search = $this->requester->get('search');
         if (!isset($search)) {
             $this->log()->error('opdsBySearch called without search criteria, page ' . $page);
             // 400 Bad request
-            return $this->responder->mkError(400);
+            return $this->responder->error(400);
         }
         $filter = $this->getFilter();
         $tl = $this->calibre()->titlesSlice($settings['lang'], $page, $settings->page_size, $filter, $search);
@@ -393,7 +393,7 @@ class OpdsActions extends DefaultActions
             $tl['total'],
             $settings->page_size
         );
-        return $this->responder->mkOpdsResponse($cat, OpdsGenerator::OPDS_MIME_ACQ);
+        return $this->responder->opds($cat, OpdsGenerator::OPDS_MIME_ACQ);
     }
 
     /**
@@ -415,7 +415,7 @@ class OpdsActions extends DefaultActions
                 $this->log()->info("logged out user: " . $username);
             }
         }
-        return $this->responder->mkAuthenticate($settings['appname']);
+        return $this->responder->authenticate($settings['appname']);
     }
 
     /*********************************************************************